Disable 'done' button
I've just realised that the user could always click on the 'done' button, however little information was entered in the wizard. I think that disabling the 'done' button for the first wizard tab and reenabling it only on the second tab is the best solution, so we make sure that the application has a minimal set of information.